trying to clean up some possible problems if there is an error on an inv dataset in the cache generation, TODO: stating using with

+++ src/pywms/grid_init_script.py
@@ -279,7 +279,16 @@ def create_topology(datasetname, url):
         logger.error("Disabling Error: " +\
                                  repr(traceback.format_exception(exc_type, exc_value,
                                               exc_traceback)))
-        os.unlink(nclocalpath)
+        try:
+            nclocal.close()
+        except:
+            pass
+        try:
+            nc.close()
+        except:
+            pass 
+        if os.path.exists(nclocalpath):
+            os.unlink(nclocalpath)
 
 def create_topology_from_config():
     """
